move all alert scheduling code into a single package. finalize.
