+++ test/java/com/arpnetworking/metrics/portal/alerts/AlertExecutionContextTest.java
@@ -17,6 +17,7 @@
 package com.arpnetworking.metrics.portal.alerts;
 
 import com.arpnetworking.metrics.portal.TestBeanFactory;
+import com.arpnetworking.metrics.portal.alerts.scheduling.AlertExecutionContext;
 import com.arpnetworking.metrics.portal.scheduling.Schedule;
 import com.arpnetworking.metrics.portal.scheduling.impl.NeverSchedule;
 import models.internal.MetricsQueryFormat;
